make sure we use full descriptions, pick up correct unit name
Matteo, this version actually works for me, up to the point where I see the
sleeps running on Stampede and get the task states back as DONE, so the full
circle!  It only fails later when trying the SMTP connection, which is not
surprising I guess.  I also used the 0.35 radical stack, so hope you can
reproduce it...  Lets see :)

The test script should finish with:

 ---------- all tasks are final
 ---------- list sessions, dump this session
<Response [200]>
{u'paths': [u'ssid.016682.0040'], u'success': True}
{u'result': {u'tasks': {u'task.000000': {u'state': u'Done'},
                        u'task.000001': {u'state': u'Done'},
                        u'task.000002': {u'state': u'Done'},
                        u'task.000003': {u'state': u'Done'},
                        u'task.000004': {u'state': u'Done'},
                        u'task.000005': {u'state': u'Done'}}},
 u'ssid': u'ssid.016682.0040',
 u'success': True}