change example to do timing, data staging
The tasks used now are:

  cp /etc/passwd -> unit_sandbox/passwd
  wc -l passed 10

The '10' is added as argument so that emgr has a runtime estimate for the unit
-- it needs that for planning.  That causes the unit to fail (no file '10'), but
otherwise the data staging and 'wc' work correctly.

The script now also times some of the calls, and prints some rates (units/sec)
for submission and state checking.  This is obviously totally dependent on
network interactions toward the rest interface.